Try save NFTListing during user query
