+++ backend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "alephium-nft-backend",
-  "version": "0.4.44",
+  "version": "0.4.45",
   "private": true,
   "scripts": {
     "build": "npx tsc",
