+++ frontend/package.json
@@ -1,6 +1,6 @@
 {
   "name": "alephium-nft-frontend",
-  "version": "0.4.44",
+  "version": "0.4.45",
   "private": true,
   "scripts": {
     "dev": "NODE_OPTIONS='--inspect' next dev -p 3020",
